fix 1st pull request changes
