+++ src/main/java/com/example/Lion.java
@@ -3,10 +3,10 @@ package com.example;
 import java.util.List;
 
 public class Lion {
-
     boolean hasMane;
 
     public Lion(String sex, Feline feline) throws Exception {
+        this.feline = feline;
         if ("Самец".equals(sex)) {
             hasMane = true;
         } else if ("Самка".equals(sex)) {
@@ -15,8 +15,7 @@ public class Lion {
             throw new Exception("Используйте допустимые значения пола животного - самец или самка");
         }
     }
-
-    Feline feline = new Feline();
+    Feline feline;
 
     public int getKittens() {
         return feline.getKittens();
