+++ src/test/java/com/example/LionParameterizedTest.java
@@ -1,0 +1,36 @@
+package com.example;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+
+import static org.junit.Assert.assertEquals;
+
+@RunWith(Parameterized.class)
+public class LionParameterizedTest {
+    Feline feline = new Feline();
+    private final String sex;
+    private final Boolean hasMane;
+
+    public LionParameterizedTest(String sex, boolean hasMane) {
+        this.sex = sex;
+        this.hasMane = hasMane;
+    }
+
+    @Parameterized.Parameters
+    public static Object[][] getSex(){
+        return new Object[][]{
+                {"Самец", true},
+                {"Самка", false}
+        };
+    }
+
+    @Test
+    public void doesHaveManeTest() throws Exception {
+        Lion lion = new Lion(sex, feline);
+        assertEquals(hasMane, lion.doesHaveMane());
+    }
+}
